% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_utils.r
\name{all_idf_filenames_contain_dates}
\alias{all_idf_filenames_contain_dates}
\title{Check if all idf filenames (like in 'HEAD_20080402_l1.idf') contain dates.
FALSE if at least 1 filename specified does not contain a date.}
\usage{
all_idf_filenames_contain_dates(idfname)
}
\arguments{
\item{idfname}{idf-filename (character).}
}
\value{
FALSE if at least 1 filename specified does not contain a date; TRUE otherwise (all filenames contain a date).
}
\description{
Check if all idf filenames (like in 'HEAD_20080402_l1.idf') contain dates.
FALSE if at least 1 filename specified does not contain a date.
}
\examples{
all_idf_filenames_contain_dates(c("HEAD_20080401_l1.idf","HEAD_20080501_l1.idf"))
}
