% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_sf-data.R
\docType{data}
\name{bacia}
\alias{bacia}
\title{Bacia hidrográfica do Rio Sousa na Senhora do Salto}
\format{An object of class \code{"sf"}.}
\usage{
data(bacia)
}
\description{
Obtida para servir de exemplo na aplicação das IDFs
}
\examples{
data(bacia)
\donttest{plot(bacia)}
}
\keyword{datasets}
