% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_idf.r
\name{create_funstr}
\alias{create_funstr}
\title{Create a string ("funstr") to be used in functions read_raster() and write_raster().}
\usage{
create_funstr(s, pattern = funstr_ptrn(), replacement = "x")
}
\arguments{
\item{s}{Blueprint of the string to be created, with the variable between square brackets.}

\item{pattern}{character string containing a regular expression.}

\item{replacement}{Symbol to replace the variable with in "s"}
}
\value{
String ("funstr") to be used in functions read_raster() and write_raster()
}
\description{
Create a string ("funstr") to be used in functions read_raster() and write_raster().
}
\examples{
replacement <- c("y","x")
create_funstr("[A]*100+[B]", replacement=c("y","x"))
}
