% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_utils.r
\name{create_statistic_raster}
\alias{create_statistic_raster}
\title{Summarize the values of multiple layers into one layer.}
\usage{
create_statistic_raster(
  x,
  EPSG = "EPSG:28992",
  e = NULL,
  funstr = NULL,
  statistic = "mean",
  keys
)
}
\arguments{
\item{x}{filename (character)}

\item{EPSG}{coordinate reference system like "EPSG:4326" (character)}

\item{e}{Terra extent object}

\item{funstr}{Optional function to modify raster values. Denote the raster with 'x' (or character between square brackets) (character)}

\item{statistic}{Statistic to apply to selected rasters: mean, min, max, sd, median (character)}

\item{keys}{a one row data.frame (or tibble) with no column, consistently with dplyr::group_keys() (tibble)}
}
\value{
terra::SpatRaster
}
\description{
Summarize the values of multiple layers into one layer.
}
\examples{
statistic <- "max"
keys <- data.frame(day=1)
f1 <- system.file("extdata", "HEAD_20080401_l1.idf", package="idf")
f2 <- system.file("extdata", "HEAD_20080501_l1.idf", package="idf")
x <- c(f1, f2)
create_statistic_raster(x, statistic=statistic, keys=keys)
}
