% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{df2i}
\alias{df2i}
\title{Obter intensidade num ponto no norte de portugal a partir da duração da chuvada e período de retorno}
\usage{
df2i(d, f, lugar)
}
\description{
Obter intensidade num ponto no norte de portugal a partir da duração da chuvada e período de retorno
}
