% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDF.R
\name{dgev.d}
\alias{dgev.d}
\title{Density function of modified generalized extreme value distribution}
\usage{
dgev.d(q, mu = 0, sigma = 1, xi = 0, theta = 0, eta = 1, d = 1,
  log = FALSE)
}
\arguments{
\item{q}{Vector of quantiles}

\item{mu}{location value}

\item{sigma}{scale value}

\item{xi}{shape value}

\item{theta}{value defining the curvature of the IDF}

\item{eta}{value defining the slope of the IDF}

\item{d}{vector of durations}

\item{log}{\code{logical} option to use logarithmic parameter values, default=FALSE}
}
\value{
dgev.d gives the density function
}
\description{
The function \code{dgev.d} is a modified version of the function \code{dgev} for different durations \code{d} developed by Koutsoyiannis et al. (1998).
}
\author{
Christoph Ritschel \email{christoph.ritschel@met.fu-berlin.de}
}
\seealso{
\code{\link[evd]{dgev}}
}

