% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevdfit.R
\docType{data}
\name{example}
\alias{example}
\title{Sampled data for duration-dependent GEV}
\format{
A data frame with 330 rows and 4 variables
}
\usage{
data('example',package ='IDF')
}
\description{
Randomly sampled data set used for running the example code, containing:
\itemize{
  \item \code{$xdat}: 'annual' maxima values
  \item \code{$ds}: corresponding durations
  \item \code{$cov1}, \code{$cov2}: covariates}
d-GEV parameters used for sampling:
\itemize{
  \item \eqn{\tilde{\mu} = 4 + 0.2 cov_1 +0.5 cov_2}
  \item \eqn{\sigma_0 = 2+0.5 cov_1}
  \item \eqn{\xi = 0.5}
  \item \eqn{\theta = 0}
  \item \eqn{\eta = 0.5}
  \item \eqn{\eta_2 = 0.5}
  \item \eqn{\tau = 0}}
}
\keyword{datasets}
