% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_utils.r
\name{filter_idfnames}
\alias{filter_idfnames}
\title{Filter filenames of idf- files with a date included, like 'HEAD_20080402_l1.idf'}
\usage{
filter_idfnames(idfname, fltr = NULL)
}
\arguments{
\item{idfname}{idf-filename (character).}

\item{fltr}{filter to be applied. In this filter a dot indicates the date field (character).}
}
\value{
Selected names of idf-files.
}
\description{
Filter filenames of idf- files with a date included, like 'HEAD_20080402_l1.idf'
}
\examples{
idfname <- c("HEAD_20080401_l1.idf","HEAD_20080501_l1")
fltr <- "filter(month==4)"
filter_idfnames(idfname, fltr)
}
