% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_utils.r
\name{funstr_ptrn}
\alias{funstr_ptrn}
\title{Create a string to be used as pattern in the function "create_funstr()".}
\usage{
funstr_ptrn()
}
\value{
string to be used as pattern in the function "create_funstr()" (character)
}
\description{
Create a string to be used as pattern in the function "create_funstr()".
}
