% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevdfit.R
\name{gev.d.init}
\alias{gev.d.init}
\title{get initial values for gev.d.fit}
\usage{
gev.d.init(xdat, ds, link)
}
\arguments{
\item{xdat}{vector of maxima for different durations}

\item{ds}{vector of durations belonging to maxima in xdat}

\item{link}{list of 5, link functions for parameters, created with \code{\link{make.link}}}
}
\value{
list of initial values for mu_tilde, sigma_0, xi, eta
}
\description{
obtain initial values by fitting every duration separately
}
\keyword{internal}
