% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDF.R
\name{IDF.nll}
\alias{IDF.nll}
\title{Negativ log-likelihood of modified GEV}
\usage{
IDF.nll(mu = 0, sigma = 1, xi = 0, theta = 0, eta = 1, x, d,
  use.log = F, DEBUG = F)
}
\arguments{
\item{mu}{location value}

\item{sigma}{scale value}

\item{xi}{shape value}

\item{theta}{value defining the curvature of the IDF}

\item{eta}{value defining the slope of the IDF}

\item{x}{vector of observations at different durations d}

\item{d}{vector of durations}

\item{use.log}{\code{logical} value for usage of logarithmic values, default is \code{FALSE}}

\item{DEBUG}{\code{logical} value for usage of debugging, if \code{TRUE} the input parameters and the value of negative
log-likelihood are printed on console.}
}
\value{
retruns weightes negative log-likelihood by number of observatons uesd
}
\description{
The function \code{IDF.nll} calculates the negative log-likelihood for a given set of model parameters
\code{mu,sigma,xi,theta,eta}, given observations \code{x} and given durations \code{d}. Options for the usage of
logartihmic values \code{use.log} and a debugging function \code{DEBUG} are available.
}
\author{
Christoph Ritschel \email{christoph.ritschel@met.fu-berlin.de}
}

