% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.r
\name{idf}
\alias{idf}
\title{idf: Extending the 'terra' package in order to be able to read and write idf rasters.}
\description{
IDF is a simple binary format used by the iMOD groundwater modelling software.
The format contains a header with grid and spatial extent information and
an array of floats.The array of floats is translated to a rectangular grid using
the ncol and nrow fields of the header.
}
\section{idf functions}{

This package extends the 'terra' package with the following functions:

\code{\link{read_raster}}

\code{\link{write_raster}}

\code{\link{create_funstr}}

\code{\link{funstr_ptrn}}

\code{\link{idfname_to_date}}

\code{\link{filter_idfnames}}

\code{\link{create_statistic_raster}}
}

