% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_sf-data.R
\docType{data}
\name{idf_sf}
\alias{idf_sf}
\title{Parâmetros das curvas IDF}
\format{An object of class \code{"tibble"}.}
\source{
\href{http://www.civil.ist.utl.pt/~mps/Mod_hid/Teoria/teoria 2007 para jpg.pdf}{Apontamentos de modelação hidrológica da Professora Maria Manuela Portela}
}
\usage{
data(idf)
}
\description{
Parâmetros das curvas IDF obtidas por \href{https://snirh.apambiente.pt/snirh/download/relatorios/relatorio_prec_intensa.pdf}{Brandão e Rodrigues 2001}
}
\examples{
data(idf)
\donttest{head(idf)}
}
\keyword{datasets}
