% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_utils.r
\name{idfname_to_date}
\alias{idfname_to_date}
\title{Extract the date part of an idf filename with a date included, like in 'HEAD_20080402_l1.idf'.}
\usage{
idfname_to_date(idfname)
}
\arguments{
\item{idfname}{idf-filename (character).}
}
\value{
Date
}
\description{
Extract the date part of an idf filename with a date included, like in 'HEAD_20080402_l1.idf'.
}
\examples{
idfname_to_date("HEAD_20080401_l1.idf")
}
