% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.R
\name{intensidade}
\alias{intensidade}
\title{Obter intensidade num ponto no norte de portugal a partir da duração da chuvada e período de retorno}
\usage{
intensidade(d, f, lugar)
}
\arguments{
\item{d}{é uma duração de classe _Duration_}

\item{f}{é um período de retorno em anos, por exemplo 20}

\item{lugar}{é um ponto no espaço de classe _sf_}
}
\description{
Obter intensidade num ponto no norte de portugal a partir da duração da chuvada e período de retorno
}
