% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf.r
\docType{package}
\name{idf}
\alias{idf}
\title{idf: Extending the 'raster' package in order to be able to read and write idf rasters.}
\description{
IDF is a simple binary format used by the iMOD groundwater modelling software.
The format contains a header with grid and spatial extent information and
an array of floats.The array of floats is translated to a rectangular grid using
the ncol and nrow fields of the header.
}
\section{idf functions}{

This package extends the 'raster' package with two functions:

\code{\link{read_raster}}

\code{\link{write_raster}}
}

