% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDF.R
\name{rgev.d}
\alias{rgev.d}
\title{Random generation for the modified generalized extreme value distribution}
\usage{
rgev.d(n, mu = 0, sigma = 1, xi = 0, theta = 0, eta = 1, d = 1)
}
\arguments{
\item{n}{Number of observations}

\item{mu}{location value}

\item{sigma}{scale value}

\item{xi}{shape value}

\item{theta}{value defining the curvature of the IDF}

\item{eta}{value defining the slope of the IDF}

\item{d}{vector of durations}
}
\value{
rgev.d generates random derivates
}
\description{
The function \code{rgev.d} is a modified version of the function \code{rgev} for different durations \code{d} developed by Koutsoyiannis et al. (1998).
}
\author{
Christoph Ritschel \email{christoph.ritschel@met.fu-berlin.de}
}
\seealso{
\code{\link[evd]{rgev}}
}

