% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_idf.r
\name{write_raster}
\alias{write_raster}
\title{Write raster data to a file.}
\usage{
write_raster(
  x,
  filename,
  e = NULL,
  constant = NULL,
  double_precision = FALSE,
  ...
)
}
\arguments{
\item{x}{terra::SpatRaster object}

\item{filename}{Output filename (character)}

\item{e}{Terra extent object}

\item{constant}{multiplication factor (numeric)}

\item{double_precision}{Write double precision idf. Only meaningfull for writing idf-files (logical)}

\item{...}{Additional arguments as in 'terra::writeRaster'.}
}
\value{
This function is used for the side-effect of writing values to a file.
}
\description{
Writes a terra::SpatRaster object to a file, using one of
the supported formats in the 'terra' package.
or an idf file format (ref. \url{https://oss.deltares.nl/web/imod}).
In that case, use the'.idf' extension in the filename.
}
